integrated validation meta data
