+++ spi/src/main/kotlin/org.opensearch.indexmanagement.spi/indexstatemanagement/model/ManagedIndexMetaData.kt
@@ -36,7 +36,7 @@ data class ManagedIndexMetaData(
     val stepMetaData: StepMetaData?,
     val policyRetryInfo: PolicyRetryInfoMetaData?,
     val info: Map<String, Any>?,
-    val validationInfo: Map<String, Any>?,
+    val validationMetaData: ValidationMetaData?,
     val id: String = NO_ID,
     val seqNo: Long = SequenceNumbers.UNASSIGNED_SEQ_NO,
     val primaryTerm: Long = SequenceNumbers.UNASSIGNED_PRIMARY_TERM
@@ -58,6 +58,7 @@ data class ManagedIndexMetaData(
         if (actionMetaData != null) resultMap[ActionMetaData.ACTION] = actionMetaData.getMapValueString()
         if (stepMetaData != null) resultMap[StepMetaData.STEP] = stepMetaData.getMapValueString()
         if (policyRetryInfo != null) resultMap[PolicyRetryInfoMetaData.RETRY_INFO] = policyRetryInfo.getMapValueString()
+        if (validationMetaData != null) resultMap[ValidationMetaData.VALIDATE] = validationMetaData.getMapValueString()
         if (info != null) resultMap[INFO] = Strings.toString(XContentFactory.jsonBuilder().map(info))
 
         return resultMap
@@ -83,6 +84,7 @@ data class ManagedIndexMetaData(
             .addObject(ActionMetaData.ACTION, actionMetaData, params, true)
             .addObject(StepMetaData.STEP, stepMetaData, params, true)
             .addObject(PolicyRetryInfoMetaData.RETRY_INFO, policyRetryInfo, params, true)
+            .addObject(ValidationMetaData.VALIDATE, validationMetaData, params, true)
             .field(INFO, info)
             .endObject()
             .endObject()
@@ -127,6 +129,7 @@ data class ManagedIndexMetaData(
             builder.addObject(StateMetaData.STATE, stateMetaData, params)
                 .addObject(ActionMetaData.ACTION, actionMetaData, params)
                 .addObject(StepMetaData.STEP, stepMetaData, params)
+                .addObject(ValidationMetaData.VALIDATE, validationMetaData, params)
         }
         builder.addObject(PolicyRetryInfoMetaData.RETRY_INFO, policyRetryInfo, params)
 
@@ -150,19 +153,13 @@ data class ManagedIndexMetaData(
         streamOutput.writeOptionalWriteable(actionMetaData)
         streamOutput.writeOptionalWriteable(stepMetaData)
         streamOutput.writeOptionalWriteable(policyRetryInfo)
+        streamOutput.writeOptionalWriteable(validationMetaData)
 
         if (info == null) {
             streamOutput.writeBoolean(false)
         } else {
             streamOutput.writeBoolean(true)
             streamOutput.writeMap(info)
-        }
-
-        if (validationInfo == null) {
-            streamOutput.writeBoolean(false)
-        } else {
-            streamOutput.writeBoolean(true)
-            streamOutput.writeMap(validationInfo)
         }
     }
 
@@ -201,14 +198,9 @@ data class ManagedIndexMetaData(
             val action: ActionMetaData? = si.readOptionalWriteable { ActionMetaData.fromStreamInput(it) }
             val step: StepMetaData? = si.readOptionalWriteable { StepMetaData.fromStreamInput(it) }
             val retryInfo: PolicyRetryInfoMetaData? = si.readOptionalWriteable { PolicyRetryInfoMetaData.fromStreamInput(it) }
+            val validate: ValidationMetaData? = si.readOptionalWriteable { ValidationMetaData.fromStreamInput(it) }
 
             val info = if (si.readBoolean()) {
-                si.readMap()
-            } else {
-                null
-            }
-
-            val validationInfo = if (si.readBoolean()) {
                 si.readMap()
             } else {
                 null
@@ -227,9 +219,9 @@ data class ManagedIndexMetaData(
                 stateMetaData = state,
                 actionMetaData = action,
                 stepMetaData = step,
+                validationMetaData = validate,
                 policyRetryInfo = retryInfo,
-                info = info,
-                validationInfo = validationInfo
+                info = info
             )
         }
 
@@ -257,9 +249,8 @@ data class ManagedIndexMetaData(
             var action: ActionMetaData? = null
             var step: StepMetaData? = null
             var retryInfo: PolicyRetryInfoMetaData? = null
-
+            var validate: ValidationMetaData? = null
             var info: Map<String, Any>? = null
-            var validationInfo: Map<String, Any>? = null
 
             XContentParserUtils.ensureExpectedToken(XContentParser.Token.START_OBJECT, xcp.currentToken(), xcp)
             while (xcp.nextToken() != XContentParser.Token.END_OBJECT) {
@@ -285,6 +276,9 @@ data class ManagedIndexMetaData(
                     StepMetaData.STEP -> {
                         step = if (xcp.currentToken() == XContentParser.Token.VALUE_NULL) null else StepMetaData.parse(xcp)
                     }
+                    ValidationMetaData.VALIDATE -> {
+                        validate = if (xcp.currentToken() == XContentParser.Token.VALUE_NULL) null else ValidationMetaData.parse(xcp)
+                    }
                     PolicyRetryInfoMetaData.RETRY_INFO -> {
                         retryInfo = PolicyRetryInfoMetaData.parse(xcp)
                     }
@@ -309,7 +303,7 @@ data class ManagedIndexMetaData(
                 step,
                 retryInfo,
                 info,
-                validationInfo,
+                validate,
                 id,
                 seqNo,
                 primaryTerm
@@ -347,9 +341,9 @@ data class ManagedIndexMetaData(
                 stateMetaData = StateMetaData.fromManagedIndexMetaDataMap(map),
                 actionMetaData = ActionMetaData.fromManagedIndexMetaDataMap(map),
                 stepMetaData = StepMetaData.fromManagedIndexMetaDataMap(map),
+                validationMetaData = ValidationMetaData.fromManagedIndexMetaDataMap(map),
                 policyRetryInfo = PolicyRetryInfoMetaData.fromManagedIndexMetaDataMap(map),
-                info = map[INFO]?.let { XContentHelper.convertToMap(JsonXContent.jsonXContent, it, false) },
-                validationInfo = map[VALIDATION_INFO]?.let { XContentHelper.convertToMap(JsonXContent.jsonXContent, it, false) },
+                info = map[INFO]?.let { XContentHelper.convertToMap(JsonXContent.jsonXContent, it, false) }
             )
         }
     }
