+++ src/test/kotlin/org/opensearch/indexmanagement/indexstatemanagement/model/ManagedIndexMetaDataTests.kt
@@ -35,9 +35,9 @@ class ManagedIndexMetaDataTests : OpenSearchTestCase() {
             stateMetaData = StateMetaData("close-index", 1234),
             actionMetaData = null,
             stepMetaData = null,
+            validationMetaData = null,
             policyRetryInfo = PolicyRetryInfoMetaData(false, 0),
-            info = mapOf("message" to "Successfully initialized policy: close_policy"),
-            validationInfo = null
+            info = mapOf("message" to "Successfully initialized policy: close_policy")
         )
 
         roundTripManagedIndexMetaData(expectedManagedIndexMetaData)
@@ -57,9 +57,9 @@ class ManagedIndexMetaDataTests : OpenSearchTestCase() {
             stateMetaData = StateMetaData("close-index", 1234),
             actionMetaData = ActionMetaData("close", 4321, 0, false, 0, 0, null),
             stepMetaData = null,
+            validationMetaData = null,
             policyRetryInfo = PolicyRetryInfoMetaData(false, 0),
-            info = mapOf("message" to "Successfully closed index"),
-            validationInfo = null
+            info = mapOf("message" to "Successfully closed index")
         )
 
         roundTripManagedIndexMetaData(expectedManagedIndexMetaData)
@@ -79,9 +79,9 @@ class ManagedIndexMetaDataTests : OpenSearchTestCase() {
             stateMetaData = StateMetaData("close-index", 1234),
             actionMetaData = ActionMetaData("close", 4321, 0, false, 0, 0, ActionProperties(3)),
             stepMetaData = null,
+            validationMetaData = null,
             policyRetryInfo = PolicyRetryInfoMetaData(false, 0),
-            info = mapOf("message" to "Successfully closed index"),
-            validationInfo = null
+            info = mapOf("message" to "Successfully closed index")
         )
 
         roundTripManagedIndexMetaData(expectedManagedIndexMetaData)
@@ -101,9 +101,9 @@ class ManagedIndexMetaDataTests : OpenSearchTestCase() {
             stateMetaData = StateMetaData("rollover-index", 1234),
             actionMetaData = ActionMetaData("rollover", 4321, 0, false, 0, 0, null),
             stepMetaData = StepMetaData("attempt_rollover", 6789, Step.StepStatus.FAILED),
+            validationMetaData = null,
             policyRetryInfo = PolicyRetryInfoMetaData(false, 0),
-            info = mapOf("message" to "There is no valid rollover_alias=null set on movies"),
-            validationInfo = null
+            info = mapOf("message" to "There is no valid rollover_alias=null set on movies")
         )
 
         roundTripManagedIndexMetaData(expectedManagedIndexMetaData)
